Clean up LangChain
