+++ tests/langchain/test_langchain_tracer.py
@@ -30,8 +30,8 @@ from mlflow.entities import Trace
 from mlflow.entities.span_event import SpanEvent
 from mlflow.entities.span_status import SpanStatus, SpanStatusCode
 from mlflow.exceptions import MlflowException
-from mlflow.langchain import _LangChainModelWrapper
 from mlflow.langchain.langchain_tracer import MlflowLangchainTracer
+from mlflow.langchain.model import _LangChainModelWrapper
 from mlflow.tracing.constant import SpanAttributeKey
 from mlflow.tracing.provider import trace_disabled
 from mlflow.utils import IS_PYDANTIC_V2_OR_NEWER
